glibc: Fix zoneinfo by installing system version
Solves a problem referenced in Linuxbrew/brew#92 where `strftime` did
not understand timezones such as `Australia/Darwin`. The formula now
creates a symbolic link in the installed glibc's share directory to a
system installed copy of the zoneinfo. If the system has not installed
zoneinfo in the default location then we do not do link anything in. A
user would have to do this step manually.

Fixes Linuxbrew/brew#92.
Closes Linuxbrew/homebrew-core#756.

Signed-off-by: Shaun Jackman <sjackman@gmail.com>