Revert "flake8: update 3.0.4 bottle for Linuxbrew."
This reverts commit d88156868ed85ebb1441ca70725a9375272b255b.

For reasons unknown, the virtualenv included in the bottle is incomplete.
It's missing, among other things, __future__.py, which leads to import failures,
making the test fail and preventing the application from running.

Note that if you install from source, the test passes. This is only
a problem with the bottle.