intltool: use env perl (Linuxbrew)
Main change: use #!/usr/bin/env perl

On some ancient platforms, /usr/bin/env might be able to handle
1 argument only. In such a case, 'perl -w' will be interpreted
as a single argument and 'env' will fail to execute it because
no file or script 'perl -w' exists on a normal platform. So, we
have to use a (better) `use warnings;` approach. We use
`inreplace` and exploit the fact that all intltool-* scripts use 
strict module.