sratoolkit: Resolve pacbio-load compile issue for Linux
After talking with the Linuxbrew maintainers it seems that some changes were not migrated across into the new Linuxbrew core repo. This version deparallelize's the build process and performs a replace in pacbio-load to ensure it can compile.

All altered code has had a conditional mac clause added as this was only seen in Linux based systems

Closes Linuxbrew/homebrew-core#5616.

Signed-off-by: Michka Popoff <michkapopoff@gmail.com>