pulseaudio: add a dependency on perl
To fix #7914, add a dependency on perl in the formula
so that XML::Parser uses a proper glibc.
Add a specification --with-database=simple to configure,
because perl depends on gdbm,
and if the dependency of pulseaudio on perl is build-time only,
pulseaudio detects and links gdbm at build-time, but cannot locate it at run-time.

Closes Linuxbrew/homebrew-core#8148.

Signed-off-by: Michka Popoff <michkapopoff@gmail.com>